ARM: dts: msm: Use psci based entry method for idle-states in SM6150
Set "entry-method" to "psci" for idle-states in SM6150.

If PSCI is detected at the boot time, kernel sets "entry-method"
property to "psci". But if no such thing is detected at boot
time, kernel will fallback to default values based on architecture.

To avoid ambiguity, it is better to always specify property in
devicetree.

Change-Id: I765ae559c94773c57a61cecc6189626e94e1a381
Signed-off-by: Asit Shah <quic_asitshah@quicinc.com>